Completely replaced SDL by SFML.
The SDL was replaced by SFML which provides a more lightweight C++
interface. To avoid problems with multithreading, the rendering thread
was removed, graphics are handled in the main loop. This also made the
EngineManager class more or less obsolete, which a little redundant,
too. The physics thread is now also handled by SFML.