Fully integrated universe cells to physics and visualisation.
A big part included modifying the bounding boxes to handle cells. There
are to modes for overlap test. One is to test for overlap for any cell,
the other defines a limit of how many cells the shapes/objects may be
apart. The first applies for the camera which has to be tested for all
cells. The second applies for collision detection, since object might
only have a maximum size of one cell.

Additionally, visuals are tested against the objects bounding box first
and against shape bounding boxes secondly. Maybe this helps for complex,
concatenated shapes.