Remove unneccessary class for writable commands
Commands with write access are wrapped by a special command which
allows for queued processing within a lambda function. This lambda
is then stored in a normal CCommand class like the command with
only read access. Previously, this was a special CCommandWritable
which only differed in a void return value.